public/requirements.txt
Django==3.2.12
django-webpack-loader==0.7.0
django-simple-menu
Jinja2
django-jinja==2.10.0
Pillow==9.0.1
django-vanilla-views==3.0.0
sorl-thumbnail==12.7.0
django-choices==1.7.2
django-environ
